// frontend/src/sections/prescriptions/detail/insurance.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../../api/graphql";

export type InsuranceProvidersVariables = {
  storeId: string;
};

export type InsuranceProvidersResult = {
  insuranceProviders: ({
  nodes: Array<{
  id: string;
}>;
});
};

export const InsuranceProviders = {
  query: "query insuranceProviders($storeId: String!) {\n  insuranceProviders(storeId: $storeId) {\n    ... on InsuranceProvidersConnector {\n      nodes {\n        id\n      }\n    }\n  }\n}",
} as TypedDocument<InsuranceProvidersResult, InsuranceProvidersVariables>;

export type InsurancePoliciesVariables = {
  storeId: string;
  nameId: string;
};

export type InsurancePoliciesResult = {
  insurancePolicies: ({
  nodes: Array<{
  id: string;
  policyNumber: string;
  discountPercentage: number;
  isActive: boolean;
  expiryDate: string;
  insuranceProviders: {
  providerName: string;
} | null;
}>;
});
};

export const InsurancePolicies = {
  query: "query insurancePolicies($storeId: String!, $nameId: String!) {\n  insurancePolicies(storeId: $storeId, nameId: $nameId) {\n    ... on InsuranceConnector {\n      nodes {\n        id\n        policyNumber\n        discountPercentage\n        isActive\n        expiryDate\n        insuranceProviders {\n          providerName\n        }\n      }\n    }\n  }\n}",
} as TypedDocument<InsurancePoliciesResult, InsurancePoliciesVariables>;